Rename `SetPredicate` to `ContainsPredicate`
Renames the `SetPredicate` to a `ContainsPredicate` to better describe
what's happening, and replaces the lookup with a linear search so that
only `Eq` is required of the predicate item. Also add two
specializations for doing contains checks on types that implement `Ord`
or `Hash + Eq`. The specializations can be handled by trait
specialization when Rust supports it in stable.

Signed-off-by: Nick Stevens <nick@bitcurry.com>