chore: Make `find_case` object safe
Apparently, a provided function can't downcast `&self` but I can if I
spread the implementation around.

Overall, trying to meet the goal of making it simple to write a
Predicate.  Anyone besides us will need to copy `default_find_case` if
they want the reference.  Not sure what would be a good way to make that
good enough for being stable.