feat: add similar-based diffing
Another feature-gated PoC to replace the `difference` crate.

This commit exposes the same interface as the `dissimilar` crate but using the more feature-packed
`similar` crate. The "difference" metric is a bit different due to `similar` supporting a `Replaced`
op in addition to `Same/Add/Remove`.

In the next commit I'm going to expose more `similar` features, such as algorithm selection and diff
ratio.

The doc-based unittests feel a bit spartan for the purpose of comparring diff crates, might need to
expand them a bit or move to standalone tests.