fix(diff): Remove duplicated output
In the tree view, we already show the original and the current value, we
shouldnt show an entire Diff that is only parseable by color.

In changing this, we removed the more cosmetic atom selector.  We also
removed the edit distance, since there isn't a known case for it.  Let
us know if you needed this!

Fixes #94
Fixes #105