fix: Report user value
With some predicates, we show the user value (diff) but others do not.
This makes it more consistent.  We aim to show the user value in leaf
predicates and in major adapters.
- Some leaf predicates can't show a value (function)
- Some defer to a file on disk, so we leave it at that, in case the file
  is too big or ugly to shoe
- Simple adapters, like `trim` don't show.  So far, the only worthwhile
  one is `eq_file`.

By leaving things to the leaves, we are able to show the value after its
gone through transforms, like UTF-8, or loading a file path.

Fixes #116